move code to the .cpp it belongs to
